third_party/WebKit/LayoutTests/http/tests/serviceworker/skip-waiting.html

<!DOCTYPE html>
<title>Service Worker: Skip waiting</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>

promise_test(function(t) {
    var scope = 'resources/simple.html';
    var url = 'resources/skip-waiting-worker.js';
    var frame, frame_sw;

    return service_worker_unregister(t, scope)
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(f) {
          frame = f;
          frame_sw = f.contentWindow.navigator.serviceWorker;
          assert_equals(frame_sw.controller, null,
                        'Document controller should be null');
          return navigator.serviceWorker.register(url, {scope: scope});
        })
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() {
          assert_equals(frame_sw.controller, null,
                        'Document controller should still be null');
        })
      .then(function() {
          frame.remove();
          return service_worker_unregister_and_done(t, scope);
        });
  }, 'Test skipWaiting while a client is not being controlled');

promise_test(function(t) {
    var scope = 'resources/blank.html';
    var url1 = 'resources/empty.js';
    var url2 = 'resources/skip-waiting-worker.js';
    var frame, frame_sw, sw_registration, oncontrollerchanged;
    var saw_controllerchanged = new Promise(function(resolve) {
        oncontrollerchanged = function(e) {
            assert_equals(e.type, 'controllerchange',
                          'Event name should be "controllerchange"');
            assert_true(
                e.target instanceof frame.contentWindow.ServiceWorkerContainer,
                'Event target should be a ServiceWorkerContainer');
            assert_equals(e.target.controller.state, 'activating',
                          'Controller state should be activating');
            assert_equals(
                frame_sw.controller.scriptURL, normalizeURL(url2),
                'Controller scriptURL should change to the second one');
            resolve();
        };
      });
    return service_worker_unregister_and_register(t, url1, scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(f) {
          frame = f;
          frame_sw = f.contentWindow.navigator.serviceWorker;
          assert_equals(
              frame_sw.controller.scriptURL, normalizeURL(url1),
              'Document controller scriptURL should equal to the first one');
          frame_sw.oncontrollerchange = t.step_func(oncontrollerchanged);
          return navigator.serviceWorker.register(url2, {scope: scope});
        })
      .then(function(registration) {
          sw_registration = registration;
          add_completion_callback(function() {
              registration.unregister();
            });
          return saw_controllerchanged;
        })
      .then(function() {
          frame.remove();
          assert_not_equals(sw_registration.active, null,
                            'Registration active worker should not be null');
          fetch_tests_from_worker(sw_registration.active);
        });
  }, 'Test skipWaiting while a client is using the registration');

promise_test(function(t) {
    var scope = 'resources/other.html';
    var url1 = 'resources/empty.js';
    var url2 = 'resources/empty-worker.js';
    var url3 = 'resources/skip-waiting-worker.js';
    var frame, sw_registration, activated_worker, waiting_worker;
    return service_worker_unregister_and_register(t, url1, scope)
      .then(function(registration) {
          sw_registration = registration;
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(f) {
          frame = f;
          return navigator.serviceWorker.register(url2, {scope: scope});
        })
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'installed');
        })
      .then(function() {
          activated_worker = sw_registration.active;
          waiting_worker = sw_registration.waiting;
          assert_equals(activated_worker.scriptURL, normalizeURL(url1),
                        'Worker with url1 should be activated');
          assert_equals(waiting_worker.scriptURL, normalizeURL(url2),
                        'Worker with url2 should be waiting');
          return navigator.serviceWorker.register(url3, {scope: scope});
        })
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() {
          assert_equals(activated_worker.state, 'redundant',
                        'Worker with url1 should be redundant');
          assert_equals(waiting_worker.state, 'redundant',
                        'Worker with url2 should be redundant');
          assert_equals(sw_registration.active.scriptURL, normalizeURL(url3),
                        'Worker with url3 should be activated');
          frame.remove();
          return service_worker_unregister_and_done(t, scope);
        });
  }, 'Test skipWaiting with both active and waiting workers');

</script>
